+++ packages/@rubriclab/ui/tailwind.config.ts
@@ -1,0 +1,17 @@
+import {Config} from 'tailwindcss'
+
+import config from '@rubriclab/tailwind-config'
+
+const tailwindConfig = {
+	content: ['./*.tsx'],
+	presets: [config],
+	theme: {
+		extend: {
+			colors: {
+				...config.theme.extend.colors
+			}
+		}
+	}
+} satisfies Config
+
+export default tailwindConfig