+++ packages/create-rubric-app/templates/fullstack/app/layout.tsx
@@ -1,5 +1,4 @@
 import ToastProvider from '@rubriclab/ui/ToastProvider'
-import '@rubriclab/ui/globals.css'
 import localFont from 'next/font/local'
 import BackgroundGrid from '../components/BackgroundGrid'
 import {META} from '../constants/metadata'
@@ -23,7 +22,7 @@ export const metadata = {
 export default function RootLayout({children}: {children: React.ReactNode}) {
 	return (
 		<html lang='en'>
-			<body className={`${calSans.variable}` + `relative flex h-full min-h-screen w-full flex-col items-center`}>
+			<body className={`${calSans.variable} relative flex h-full min-h-screen w-full flex-col items-center`}>
 				<BackgroundGrid className='fixed h-full w-full' />
 				<ToastProvider />
 				<div className='z-10 w-full max-w-6xl p-5 pt-0'>{children}</div>